refactor: organize events into dedicated module
Move VM0 event-related code into a dedicated events module for
better organization and maintainability.

Changes:
- Create src/lib/events/ directory
- Move vm0-events.ts -> events/vm0-events.ts
- Move types/vm0-events.ts -> events/types.ts
- Add events/index.ts for unified exports
- Update all imports to use new events module

Structure:
src/lib/events/
├── types.ts        (event type definitions)
├── vm0-events.ts   (VM0 event service)
└── index.ts        (unified exports)
