+++ examples/demo.yaml
@@ -1,0 +1,26 @@
+version: "1.0"
+
+agent:
+  name: vm0-git-volume-test
+  description: "Test agent with git volume"
+  image: vm0-claude-code-dev
+  provider: claude-code
+  working_dir: /home/user/workspace
+  volumes:
+    - claude-system:/home/user/.claude
+    - user-workspace:/home/user/workspace
+
+volumes:
+  claude-system:
+    driver: s3fs
+    driver_opts:
+      uri: s3://vm0-s3-demo-bucket-public/claude-files
+      region: us-west-2
+
+dynamic_volumes:
+  user-workspace:
+    driver: git
+    driver_opts:
+      uri: https://github.com/{{user}}/question.git
+      branch: main
+      token: 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
