+++ turbo/apps/web/src/lib/events/index.ts
@@ -1,0 +1,7 @@
+/**
+ * Events Module
+ * Centralized module for VM0 system events management
+ */
+
+export * from "./types";
+export * from "./vm0-events";
